requirements.txt
Flask
qrcode
pillow
gunicorn
pandas
streamlit
psycopg2-binary
SQLAlchemy
resend
pyotp